requirements.txt
anthropic>=0.39.0
python-docx>=0.8.11
openpyxl>=3.1.2
rich>=13.0.0
pandas>=2.0.0
openai>=1.0.0
python-docx>=0.8.11
openpyxl>=3.1.2
jinja2>=3.1.2
rich>=13.0.0
questionary>=2.0.0
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.18.0
PyPDF2>=3.0.0